RavenDB-21623 Move about view to separate file
